+++ lib/builtin/extensions_ui/extensions_ui.dart
@@ -1,1 +1,2 @@
 export 'src/extension.dart';
+export 'src/extensions_notice.dart';